requirements.txt
# Virtual Software Company - Next Generation Requirements
google-generativeai>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pygame>=2.5.0